fix: delete incorrect logics
